+++ housekeeping/src/main/java/com/hotels/housekeeping/fs/HousekeepingS3AFileSystem.java
@@ -1,0 +1,46 @@
+/**
+ * Copyright (C) 2016-2019 Expedia Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.hotels.housekeeping.fs;
+
+import java.io.IOException;
+
+import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.fs.s3a.S3AFileSystem;
+
+public class HousekeepingS3AFileSystem implements HousekeepingFileSystem {
+
+  private final S3AFileSystem fs;
+
+  public HousekeepingS3AFileSystem(S3AFileSystem fileSystem) {
+    fs = fileSystem;
+  }
+
+  @Override
+  public boolean exists(Path path) throws IOException {
+    return fs.exists(path);
+  }
+
+  @Override
+  public boolean delete(Path path, boolean recursive) throws IOException {
+    return fs.delete(path, recursive);
+  }
+
+  @Override
+  public boolean isEmpty(Path path) throws IOException {
+    return !fs.exists(path) || fs.getFileStatus(path).isEmptyDirectory();
+  }
+
+}